practice/practice.py
# Practicing each of the essential Python concepts mentioned earlier is crucial for gaining proficiency. Here are some exercises and challenges tailored to specific concepts:

# 1. **Basic Syntax and Variables:**
#    - Write a Python program to calculate the area of a rectangle.
#    - Create a program that swaps the values of two variables without using a temporary variable.

def calculate_area(length=0, width=0):
    try:
        l = float(length)
        w = float(width)
        if l <= 0 or w <= 0:
            return 0
        else:
            return l * w
    except (ValueError, TypeError):
        return 0



# 2. **Data Types:**
#    - Create a program that converts temperatures between Fahrenheit and Celsius.

def temp_converter(temp, type="c"):
    try:
        if temp == "":
            raise ValueError("Please enter a number for the temperature")
        temp = int(temp)
        type = str(type).lower()
        if type == "c":
            result = (temp * 9/5) + 32
            new_type = "f"
        elif type == "f":
            result = (temp - 32) * 5/9
            new_type = "c"
        else:
            raise ValueError("Please enter 'c' or 'f' for the type")
        return f"{result} {new_type}"
    except TypeError:
        raise ValueError("Please enter a number for the temperature")


#    - Write a script that combines multiple strings into a single one.

def string_combiner(a="",b=""):
    return str(a) + str(b)



# 3. **Conditional Statements:**
#    - Build a simple calculator that performs addition, subtraction, multiplication, and division based on user input.

#    - Write a program that determines whether a given number is even or odd.

# 4. **Loops:**
#    - Implement a program to print the Fibonacci sequence up to a given number.
#    - Write a script to find the factorial of a number using a `while` loop.

# 5. **Functions:**
#    - Create a function that calculates the area of a circle based on its radius.
#    - Write a function to check if a given number is prime.

# 6. **Error Handling:**
#    - Build a program that handles exceptions when dividing by zero or performing invalid operations.
#    - Create a script that reads data from a file and handles potential file-related exceptions.

# 7. **File Handling:**
#    - Write a program that reads data from a file, processes it, and writes the results to a new file.
#    - Create a script that appends data to an existing file.

# 8. **Lists and List Comprehensions:**
#    - Implement a script that finds the largest and smallest elements in a list.
#    - Use list comprehensions to generate a list of squares from 1 to 10.

# 9. **Dictionaries:**
#    - Build a program that stores information about books in a dictionary and allows users to search for books by title or author.
#    - Create a dictionary of students' grades and calculate the average grade.

# 10. **Modules and Libraries:**
#     - Use the `random` module to simulate a dice roll.
#     - Import the `math` module to calculate the square root of a number.

# 11. **Object-Oriented Programming (OOP):**
#     - Define a class representing a basic shape with methods to calculate area and perimeter for different shapes (e.g., rectangle, circle).
#     - Create a class for a simple bank account with methods to deposit, withdraw, and check the balance.

# 12. **String Manipulation:**
#     - Write a program that reverses a given string.
#     - Implement a function to count the occurrences of a specific word in a text.

# 13. **Math and Number Operations:**
#     - Build a program to calculate the sum of all even numbers in a given range.
#     - Write a script that computes the factorial of a number using a loop.

# 14. **List Sorting and Searching:**
#     - Create a script to sort a list of numbers in ascending and descending order.
#     - Implement a binary search algorithm to find an element in a sorted list.

# 15. **Recursion:**
#     - Write a recursive function to calculate the nth Fibonacci number.
#     - Create a recursive function to calculate the factorial of a number.

# 16. **Data Structures and Algorithms:**
#     - Implement a stack or queue data structure.
#     - Write sorting algorithms like Bubble Sort or Quick Sort.

# 17. **Time and Date Manipulation:**
#     - Build a program that displays the current date and time.
#     - Calculate the number of days between two given dates.

# 18. **Regular Expressions:**
#     - Create a program that validates email addresses using regular expressions.
#     - Write a script that extracts all the URLs from a text using regular expressions.

# 19. **Exception Handling:**
#     - Handle exceptions when dividing by zero, and provide appropriate error messages.
#     - Create custom exceptions and use them in your code.

# 20. **Basic Input/Output:**
#     - Build a simple command-line calculator that takes user input for calculations.
#     - Create a program with a graphical user interface (GUI) that takes user input and displays results.

# 21. **Unit Testing:**
#     - Write unit tests for some of the functions or classes you've created in the previous exercises using a testing framework like `unittest` or `pytest`.

# 22. **Debugging Techniques:**
#     - Introduce intentional bugs in your code and practice debugging them using tools like `pdb` or print statements.

# 23. **Version Control:**
#     - Set up a Git repository for your coding exercises and practice committing, branching, and merging.

# 24. **Coding Style and PEP 8:**
#     - Ensure your code follows PEP 8 guidelines for code formatting and readability.


# print(area(5,6))
# print(string_combiner())
# print(temp_converter(32, "c"))
